feat(broker): add live progress logging to the VSS self-test round trip
Every stage of --self-test-vss was previously silent between "Running:"
and the final PASS/FAIL line, so a hang (like the arg-quoting bug just
fixed) gave zero indication of where it was stuck. Adds tracing::info!
checkpoints at every step: pipe creation, helper spawn (with pid),
waiting for the helper to connect, waiting for its Ready/Failed event,
snapshot creation, marker verification, and deletion/release.

Split create_snapshot into wait_for_helper_ready + finish_create_snapshot
to keep its cognitive complexity under the lint ceiling with the new
logging in place, and split the self-test round-trip logic out of
vss_helper.rs into a new sibling file (vss_self_test.rs) since the
added instrumentation pushed vss_helper.rs over the workspace's 800-LOC
file-size policy.
